+++ numba/targets/externals.py
@@ -9,6 +9,7 @@ from llvmlite import ir
 import llvmlite.binding as ll
 
 from numba import llvmthreadsafe as llvmts
+from numba import llvmthreadsafeapi as llvmtsapi
 from numba import utils, config
 from numba import _helperlib
 from . import intrinsics
@@ -113,6 +114,7 @@ class _Installer(object):
 
     _installed = False
 
+    @llvmts.lock_llvm
     def install(self, context):
         """
         Install the functions into LLVM.  This only needs to be done once,
@@ -196,7 +198,7 @@ define void @fnclex() {
 }
     """
     ll.initialize_native_asmparser()
-    library.add_llvm_module(llvmts.parse_assembly(ir_mod))
+    library.add_llvm_module(llvmtsapi.parse_assembly(ir_mod))
     library.finalize()
     return library
